Proof of concept for use of PYVPLM on props
These notebooks have been used to get more familiar with the possiiblities offered by PyVPLM. The last one shows how an expression for the efficiency can be obtained with a fairly low order and good accuracy. The accucracy on the zones with high efficiency are good but slightly above the 5% mark, it is worth investigating how often they are reached. It should also be interesting to improve the precision of the set we use to establish the law with increased focus on those zones of high efficiency