Comparaison of map vs VPLM and varying diameter
Compared the results of a regression using PyVPLM and the results of our map to see if the results were valid and got satisfactory results when comparing to the propeller of the TBM900
Also started studying a variation of the dimater and the number of blades. The curves seem to be a bit off which might indicate a need for a similarity parameter. We will try keeping the number of blade constant for now and instead study the impact of a change in the chord and twist law