Added thermal analysis to performances
Added the computation of the influence of the inverter temperature on the resistance (not gate voltage for now), and the computation of the increase in temperature based on the losses (very simple lumped model for now). Temperature derivative and increase were coded and tested but not included in the performances computation. Some minor correction on name were made, and the reference IGBT module was changed to one of a more recent technology.