Finished Inverter performances
Finished the implementation of the modules necessary to compute the performances of the inverter. Two were made as implicit component to leave more DoF to the solver, they might be changed later. Added the inverter in the simple assembly, still working so far and the efficiencies seem consistent albeit a bit above average for the inverter (0.99 it would seem, usually rather around 0.98)