DC Bus model
Added a model for the DC bus, the performance part comes from the load flow analysis method but the sizing part is made from cable models as well as some other paper. The Assembly is not working right now as a DC/DC converter model is required and will be added next.
Also disabled reports where possible because it was starting to get on my nerves
Some small corrections as well