DC/DC converter efficiency
Added an explicit computation of the DC/DC converter efficiency which relies on the same losses computation as the IGBT modules in the inverter and a very simple topology for the buck-booster. The article that was used as a reference also uses the losses in the iductantce and capacitance but some other article on more complex topologies don't include them. Also, no dependency on the output voltage