Addition of capa and inductor losses
Added to the computation of losses the conduction losses in the capacitor and inductor. Also added a computation of the required inductance and capacitance. Theses however leads to absurd weight value which point to a need to change the weight model which were therefore not implemented. The DC/DC converter (buck-boost) is going to be left as is for now.