Simple battery model
Very cimplified battery model for the base of the code and to check compatibility with the rest of the model. Computes OCV and internal resistance and adjusts current and voltage to match the i-V curve. Regressions based on a pouch cell of 20Ah. SoC does not update yet but it will come in the next commit.