Battery model and GitHub workflow
Finished the implementation of the model for the estimation of the performances of the battery with the addition of SOC variation as well as heat losses for future TMS implementation. The model remain very simple and cell dependant !
Also added an estimation of the battery mass (again very simple, no sizing of cooling nor any other "auxiliary" components of the battery)
Added the files necessary for the GitHub worflow to run but without codacy coverage for now, as it wasn't deemed useful.