Improved thermal model for the inverter
We now consider individually the junction temperature for the diode and the IGBT swith as well as the casing temperature, instead of the inverter temperature as a whole.
We however assume that the casing is common to the 2 diodes and the two IGBT switch, which, according to SEMIKRON datasheet is a strong hypothesis (which is as valid as two "different" casings.