DC bus maxima rewriting and k-factors
Rewriting the identification of the maximum current/voltage seen by the DC/DC bus to match what was done in the cable (only one component to identify all maxima).
Added k_factors for the conductor and insulation radius to allow to have a lever of action in case the cooling isn't sufficient to maintain the cable temperature low enough.