PMSM Maxima and renaming
Added a component to identify the maxima in the PMSM and renamed some of the inputs. The convention is now that _max is linked with the maximum seen during the mission while _caliber is linked with maxium the component can actually withstand/do. That means that, for constraints to be respected _max should be less than _caliber.