Battery constraints
Contrarily to the rest of the components, the default battery constraints is for now to ensure. The enforce part is still coded but is a bit trickier. The goal of the enforce is to ensure that the SOC at the end of mission is greater than the minimum but not so high that we could remove pack.