Forth attempt
The code now runs with enforce on all components except PMSM and battery (by design on the latter). A line search algorithm was added to enforce bounds on some components (mainly losses which should always be >0 except for entropic losses), which is to be generalized on all code variables to prevent impossible state to be reached and send the convergence spiraling.
Also added a clipping of the SOC, but not using OpenMDAO's bound system in the component that computes the open-circuit voltage and internal resistance