Added battery enforcement
Tested the battery enforcement on a real case which lead to a couple of modification mainly the fact that the number of module, since it is a variable that OpenMDAO will loop on is not an integer but a float. As such when choosing the enforce, the user should always do a re-run afterward with the number rounded up and the ensure constraint on the battery.
Also changed the tests a little bit so that they run unit tests and assembly tests without writing the coverage we don't use.