Small work on propeller regression
Changed the dependency of pyvplm to the master to bypass a release error.
Also reworked on the propeller regression. The singular point for the previous formula were identified as fringe cases but it prompted the start of a work on a new regression which only consider efficiency greater than 0.5. Less term are needed for the same accuracy and it behaves well on the control point with efficiency greater than 0.5 and Okay-ish for the rest of the points.