Varying gate voltage
Implemented a model which makes the gate voltage vary with the temperature.
Minimal change on the efficiency of the inverter.
This model as well as the one of the resistance as a function of temperature are not yet implemented on the DC/DC converter which should eventually be done because the same IGBT modules are used.