Linking switching frequency DC/DC
Lined the switching frequency used for the computation of switching losses to the frequency used for the sizing of the inductor and capacitor. Did it so by using the constraints, so one way of handling the switching is by using the ivc in the performances. Will do the same in the inverter.