First attempt at reading power train files
With this commit it is possible to read a yaml file containing the the list of component of a hybrid electric power train and recognize which component it is based on a determined set of id (a  mechanic such as the regstration of id is not yet planned). Nothing can be done from it just now however.