Improvement of power traun file reading
Refined the definition of components inthe ower train file and improved the reading as well.
It now reads the type of components, their name and options.
It however appears that the building of class directly inside a function is quite complicated so instead of the building will be done in special components  which will appear in the next commit