Power train mass
The component that dos the sizing of the power train now has an additionnal subsystem which sums all the weight based on the name in the pt file. The output of sid subsystem is then reused in a submodel with a service name from FAST-OAD-GA to be at the proper spot.