Interfacing and computation time
Added the interface with the misson vector for the energy consumption and fuel consumed and since we were close to being able to do a proof of concept, tried running the code with 250 element (number of point for the mission). It either did not converge properly (with Newton, the last point failed) or create a Malloc error (no idea what it is but sounds bad, with a BroydenSolver) . Seems to strenghten the need to reduce computation may even force us to Mux/Demux.