Changed solver
According to OpenMDAO's documentation (https://openmdao.org/newdocs/versions/latest/features/building_blocks/solvers/nonlinear_block_gs.html), NonLinearBlockGS should be used in case such as the ones the components performances were in. Greatly reduce computational time. Could also be applied on Performance chain level but the thrust splitter and enrgy consumption component seem to be causing an issue for the simple assembly. Might need to isolte the performances in a group with NonLinearBlockGS and add a LinearRunOnce to the group that will envelop performances + components that do the interface