Copy Paste of FAST-GA mission
Copy pasted the mission from FAST-GA into FAST-GA-HE and removed all references to propulsion id to prepare for the interfacing with the HE powertrain. Added a dummy engine.
Some classes from FAST-GA might be reused but since the plan is also to add the number of points of each phase as an option it might not be a lot