Preping for interface
Added some fake option for the submodel that computes the energy consumption so that both a propulsion_id and a power_train_file_path can be used. This forces us to have the options declared even when not used.
However it looks like some connection are need on mission level but not used on pt file path level which creates the need for a group that engulfs the PerformancesFromPTFile and adds a "fake interface" with all required aircraft level inputs and outputs