Moved taxi out in array
Previously, the taxi out and in phase were concatenated at the end of the arrays containing the flight conditions of the mission. This was valid only for fuel burning engine, which is the reason why we change it here so that the taxi out is done before the mission.
This commit also highlight the need for a proper reserve phase computation which has the same problem, where we can't just compute it "offline".