Proper reserve implementation
Now that the performances of the power train strogly depend on the state of the energy source (SoC), a proposer implementation that did not rely on a simple "règle de trois" was necessary.
With this commit, the reserve is now a proper phase, for which the number of points can be handled. Also added the sparse derivatives when necessary and added a couple of partial derivative computation to save more time. There might be other commit to add some partials. Also added some proper components for the computation of speed in the different flight phases.