First attempt at the mock-up interface
All connexion seems to be working properly, there are however some variables that needed to be renamed during the promote. Now we are missing a component that computes the exterior temperature at all points of the flight.
The mock-up interface contains fake inputs and a computation of the "thrust_rate" which is rather the computation of the ratio of available power of PROPULSIVE LOAD