Working with mission vector
Mock-up interface is done and the power train builder is now working with the mission. Added a component for the computation of the exterior temperature at each altitude which is connected to the energy computation, meaning all submodel for this service will required it, even as a fake input. The computation time is however fairly long as is (+20 sec for 90 points in the mission)