Option for solver deactivation
Added an option to deactivate solvers in the power train performances computation. This way, when running it joitly with the mission we can save time by letting the mission solver handle the convergence. By default the solvers are added and only deactivated when called in the interface with the mission.