Remove overlap when mission and pt
We now remove the overlap of data when both the mission data and power train data are present. This is done thanks to name we promote off of the list in powertrain.py. Also removed the case where we only had mission data because it does not have much sense to keep since we already have the MissionViewer for that