Small improvement to performances watcher
Made some small tweeks on the performances watch, mainly that it takes time by default on the x_axis now. Further tweek the sample AC to get closer to a viable power train (mainly greatly improved Energy density of batt cells [backed by paper] and the mission RPM to resemble more an EMRAX engine). The cable temperature is to be checked !