Cable gauge model change
Previous model was giving erroneous cable gauge, which lead to very small section leading to numerous issues, including a cable temperature way higher than the acceptable limits. It is now change. An attempt at bringing back the old cable temperature model (which was deactivated because it could not converge) will be made in next commit