Cable temperature with dynamics
The computation of the temperature profile inside the cable is now a submodel, either a steady state analysis or an analysis with dynamic evolution.  For compatibility, this means the steady state needs the time_step even though it is not used.
The last time step of descent suffered from the same problem as the last one in climb since we now compute reserve as a proper flight point. Corrected with this commit