Inverter constraint separation and reference aircraft
Separated the constraints on the inverter so that they can be chosen manually.
Also corrected a mistake on the scaling of the coefficients in front of the current squared and the costant term in the computation of the switching losses.
Finalized the computation of a reference aircraft, which somewhat resembles a Pipistrel (though heavier), which will later be used to test convergences. Not yet the reference aircraft of the end of phases 1 of the PhD