Bounds enforcing on duty cycle
Early iterations of the code can lead to unfeasible value for the duty cycle, preventing the code from convergind (sqrt(<0) for instance). We now enforce it to be between 0 and 1. A precautionnary check after the convergence is recommended.
Also it seems OpenMDAO allows to give initial guesses at group level, might be used for convergence.