Battery CG, dimensions and drag
As a preamble, it was decided by the choices in this commit that all components should have a position option in the sizing even if the choice are simplistic. (Looking at you cables [for which the accurate position will be hard to get]).
Five position available for the battery each associated with their own dimensions computation method, their own CG computation and drag computation (might be zero when "inside" the fuselage).