Harnes CG, drag and length
Added the position option to the harness. There are 7 possible position each with their CG  and length computation. Length is now computed to reflect the change in aircraft dimensions but they do not precisely account for the source and target of the harness which could be done later on (We can image a X, Y, Z description for all components and compute cable dimensions from connected source and target).