Powertrain drag
Powertrain obtained by summin the drag of all the components.

There was a weird bug however with the IndepVarComp for the drag outputing which cause them to appear as inputs and outputs. This was solved by changing them to IndepVarComp with only one output (¯\_(ツ)_/¯).

Still not used in the sizing process