Small fix and position
Small fix on the computation of systems weight where the default value of some components was 1.0 instead on Nan so, when deactivating the submodel, the default value was taken at 1.0 instead of being "asked".

It is now possible to specifiy the position of PT component inside the powertrin file. If not specified the default position is taken.