Dependency of Rint to temperature
The impact of temperature on the internal resistance was identified on a 18650 cell, not the one we plan on using though. We will however make the assumption that it will be the same for our cell which is also a Li-ion 18650 type cell. Not implemented yet, will come with the rework of Rint = (SOC) and OCV = f(SOC)