Change of reference cell
Changed the reference cell for the battery construction, it is now the one used in the Pipistrel. Consequently changed the value for unit tests so that they wouldn'tt crash because of absurd values.
Added a couple of safeguards to avoid having intermediate that would make the code crash
Also changed some of the default value to match the new cell.