Addition of battery relative capacity
Added a component that computes the relative capacity of the battery depending on the current drawn. This has the effect of a quicker SOC discharge than it used to.
Further work on the Pipistrel PoC with more research on the actual data of the Pipistrel, stall speed, realistic mission, the fact that after reserve SOC is equal to 0.0, ....