Sizing of inductor
Implemented, in the OpenMDAO format, the sizing methodology for the inductor, which allows to get a more precise breakdown of the weight of the DC/DC converter. This weight is the greatest contributor to the DC/DC converter weight and might lead to erroneous results. Consequently, the option to use the power to mass was still left as a submodel.