Functionning SSPC module
SSPC module is now functionning with a very big singularity. For the integration in the power train builder we will need to detect if the SSPC is connected to a bus input or output (see why in the perf_voltage_out.py comments). Models will also need to be refines with efficiency computation and power losses computation. The temperature/resistance loop will not be implemented yet.