Poetry update and SSPC on quad architecture
Updated the poetry.lock to match the latest version of FAST-GA, no issue detected. Also added the SSPC to the quad propeller architecture and changed the results accordingly. We will now try to open these SSPC to see what happens. Also realized that the harness group tend to not covnerge in the first iterations so reduce the number of max iterations to save time.