SSPC "Logic" Implemented
To ensure that the computation of the performances of the architecture including SSPC converges properly, a certain number of construction rules were implemented based on the observation on previous commit. Mainly that if a cable is connected to a SSPC it must be connected to 2, and that if those two SSPCs are in a different state, they will be forced at the open state.