Test configuration with two converters
Because the splitter is expected to work like a DC bus (it is actually going to be a specific type of bus with 2 inputs and one outputs), I wanted to try and see if the code would converge on a bus with two inputs with DC/DC converters of different voltage output, which, unsurprisingly is not the case. This test proves it ! This means another "logic" will need to be implemented, which check that two converters on the same bus should have the same target_output_voltage.