Perforamnces of DC Splitter
First attempt at the implementation of the performances of a DC "Splitter". It was at first meant to be used as an addition to a bus, but it will actually be a special type of bus. Meaning current in and out are inputs and voltage are outputs. The way it is implemented for now gives proper results but it adds quite a few iteration to converge. Some empty components were added for weight but the same sizing procedure as buses will be applied in future commits