Attempt at SSPC in full sizing process
Attempted at adding and SSPC in the power traint for the full sizing of an aircraft. This showed a problem with the code. If the ensure constraints is chosen for the SSPC and if the value of current caliber is too low compared to what actually goes through the SSPC the code will simply stop adavancing (not a crash, just not doing any progress anymore). This means that the enforce constraints must be taken by default (will be changed in next commit). This was however deemed dangerous for inactive SSPC, so to address both issue, the enforce will be taken as a default but it will provide a minimum value in case SSPC is inactive.