Full sizing achieved
The code works in integration tests and in the notebook as well but it does highlight something very impotant, which is that there is a very strong sensitivity to the initial point given in the add_output of the components. Which means that these won't be usable for high power architecture and low power architecture, without their default value being changed. The robustness of the code must be worked on.