Rectifier performances
Implementation of the bare bones of the rectifier performances. Voluntarily simple (the plan is to use the same losses formula as for the inverters, more or less), this should simply contain a simple working model to quickly advance to fuel burning components. Value of efficiency and soon to be power density taken from Pettes-Duler's thesis