Folder setup and Rectifier correction
Set up of the folder for the generator.
Also renamed the inputs/outputs variables for the rectifier and their number to match the name at the input of the motor. Their type however, do not match, it should be investigated as mentioned in #16
Also moved the responsibility of the computation voltage_peak_in to the generator.