Sizing Generator
Added models for the sizing of the generator. Again, does not differ much from the PMSM models except for the possible positions, only in the back or in the nose,. Drag will always be zero. Also corrected a error in the phase resistance model. To be tested in an assembly and added to the powertrain builder next.