Rectifier and Generator working
Simple test on a configuration not using the powertrain builder for now show that the evaluation of the performances of the rectifier and generator work well enough. There is however yet another error (corrected as of now) which show the importance of having an initial guess which is not absurd. Reinforce the need for a preconditioning !