Integration in the PT builder
The ICE now works well with the PT builder. A couple of sanity checks will follow: mainly whether the code is capable of handling to energy source of different type or not and whether we can converge to more or less the same sizing on a classical configuration (propeller and ICE only) as with the original code.