OAD process working for full fuel
The OAD process seems to be working for the full fuel configuration. The search to make it work however highlightted one again the importance of starting point. Since the CG was originally computed after the performances, the cg variation in flight gave unplausible results, making the code crash. CG computation was thus moved above the mission and now it converges properly. There are still some slight differences with the old OAD process, mainly about the full consumption. The implementation of the effective propeller efficiency would be a starting point.
Also now big change in the powertrain builder. Instead of checking for equality of component type we chack if the type is in the list of component types. This allow for instance for the ICE to be both a source and a propulsive load.