Power Share working above threshold
A new splitter mode was added to enable to switch from a mode where the power is distributed according to a percentage to a mode where the power is shared according to a threshold value. As of now this power splitter works as long as the power output is above the threshold, if below, it will remain at the threshold value which overestimates the energy consumption of the primiary branch.