Power Split accelerated
Extended a bit the scope of what was tested on the power split mainly added two test where at least part of the split was at 100% and realized how slow it converged. To speed it up, I added a fudge factor which increase the residuals when the current in the is not close enough to zero when it should. It looks like it works by trial and error.