Power share working in all situations
It now seems that the power share mode works regardless of whether the power share is above, below or sometimes and sometimes below the output power. This was done by finding the equivalent power split (and clipping above 100) and also some shenanigans on the partials computation. Next commit is to try defining it using the pt builder