Power share mode integrated in PT builder
The power share mode is now properly working with the power train builder which entailed a few changes to the components, mainly giving up the tuning parameter that were added in the partial derivatives computation of the electric node. Instead some cutoff parameter were added which might serve as tuning parameter to filter out convergence residuals who should be equal to 0. New thing to keep in mind when assembling component. Always test them with the SSPC at the end. Final warning, in the percent split equivalent, derivative that were supposed to be 0.0 at times were put at a very smal value instead to prevent the solver to consider the jacobian as not full rank. Remains a test of the full sizig of an aircraft with the power sharer.