Simplifed AC bus
Added a simplified AC bus in order to attempt to solve the convergence problem there is on the turbo branch. We will atempt to put an AC bus between the generator and the rectifier and see if it works. This means the generator will be change a bit, see next commit