Residuals viewer
Added a function that automatically plots the residuals of a problem based on the list, for each component of what to plot. This list of variables is in the registered_components.py file.
The path to a recoder data must be provided, as well as the pt_file that was used to compute the mission. For now this only applies to run on the MissionVector group !