Implemented test on voltage coherence
A subgoal of this branch was to also implement a check that ensures that the target voltage given for voltage setters were coherent when on the same subgraphs. The reason behind that is that we know from experience that if thay are not the code will not converge, so might as well save use some time. This is done with this commit