Installed efficiency
Added a simple factor for installed efficiency, can be changed by the user, but the idea is that it represents both the effective advance ratio but also the decreased efficiency which represents increase profile drag of the cowling. The default value is based on classes at Supaero, but more detailed models would be envisioned.

Since this PR will change a lot of the values, we will not change Mission Tests and Assembly tests until the end.