Proper check for performances watcher
There was an issue with the new location of the performances watcher where it would cause the process to crash when a mission vector was launched with the "old" subsystem for energy consumption. A proper check on which submodel to use was implement, and a warning is logged when the "old" model is used to make sure the user realize he is not going to save the data from the performance watcher.