Breaking the loop in harness model
Previously there was a loop in the harness model because of the fact that resistance depended on temperature, which depended on losses, which dependend on resistance. Turns out, with the steady-state temperature model, it is easy enough to manipulate the equation so that the resistance is straight up computed, thus breaking the loop. So a new submodel was added but because it is only valid under the assumption of steady state temperature profile, a warning is raised if the wrong temperature submodel is used.
This new implementation works in the sens that now the harness systematically converges in 2 evaluation of the solver but paradoxically, on the test case it increases the computation time because the model is called more times (printed self.iter_count_apply)