Temperature profile added
Added the iterations on the temperature profiles.
Made it into a subgroup because I don't want to iterate on the relation from the load flow analysisonly on the temperature inside the group. The iteration on the equations from the load flow analysis will come from the interactions with the other components