Delta Cd propeller
Addition of a component that computes the increase in drag from the propeller. As is explained in the docstring since the increase in lift induced drag depends on the square of the lift, the individual contribution can't be summed simply. Insteas we will mark the increase in lift that contribute to the wing lift wit h a special tag and an assembler will add them to compute the global lift induced drag