Correction and verification
The implementation of the model was wrong, some variable were not promoted and hence taken at their default value which was absurd. Also revert one of the choice I took on assessing the contraction ratio at the wing AC instead of the LE which overestimated an already big result. Implemented a small verification test to compare with litterature