Mission test actually passing
Tried the mission with the new computation of the slipstream effects. It is working and is somewhat affecting the fuel consumed during the mission albeit very little. A few things to be wary of in declare_partials val=1.0 != val=np.eye(number_of_points), the solver seems to nudge the solution within the tolerance when it doesn't evolve enough. Never, ever, start a sum by doing total_cd = inputs["delta_Cdi"] because it affects the object, not the value !!!