All mission tests passing
All mission tests hav been modified to use the new module for the computation of slipstream effects, and the values have been changed accordingly. Not too much difference right now but I suspect it is because I use propellers with large diameter and a very big wing (so a small Cl so a small delta_Cl). It is worth noting that this new module significantly increases computational time