Prep for sizing wing equilibrium
Started copying and adapting the original method from fast-oad-cs23 with the new model but realized something about the Preparation for energy consumption. unfortunately, as of now, it requires the number of point in each phase which is unpractical and requires the DEP equilibrium to have as option the number in each phase instead of only the total number. Will be reworked in next commit