Works with complex propulsion
It seems to be now working with the more complex models for propulsion and DEP effect although it takes 3x longer. A few things to correct. As of now the power rate is computed based on the max value and not the caliber value so it will always be 1 ... to correct