Added test with DEP conf
Added a small test file with a DEP-specialized powertrain, a few remarks from my experimentation. To truly gain from DEP you must really cover as much span as possible with really small propeller, else the gain are not significant. The code takes too long to run when there are too many components. We could really use spoilers to increase the thrust we have to produce to reach equilibrium (and thus increase the delta_Cl). Thrust is really small at low speed (even with a Cd of 0.15)