Constant efficiency and temp inverter
Added two new submodels that renders the junction temperature and efficiency of the converter constant. One must however be careful with those model since, for now, fixing some value does not means some other value are not computed. For instance here, fixing the efficiency does not fix the losses accordingly. It also means that we could save time by deactivating those losses. However, doing so would lead to issue since losses are always promoted to the performances watcher and to the maximum components. The latter can be dealt with by putting maximums where the value are computed but for the former something needs to be devised.