Preparation for distributed loads battery
Added a component that does a preparation for the computation of the influence of the battery on the wing loads. This template should be used for any distributed loads. For punctual mass, even though the Y ratio is required, the absolute Y will be computed. Finally, the distributed tag will be changed so that it contains the list of position for which the components is to be considered distributed. Also, a field containing the list of position for which the componentn is to be considered attached to the wing will be required. We will considered that if the position is not among the one listed, it will be on the fuselage