Backpedal on handling loads
Changed my mind on how to handle the recognition of loads on the wings. This will be less elegant but easier to understand. Basically, each component defines the positions for which he should be considered as a distributed load and as punctual load for the wing. Same will be done for the fuselage later on.