Identification of punctual and distributed masses
Added two function in the power train configurator that lists the punctual and distributed masses. However, this doesn't consider the fact that, in situation where there are 2 motor+propellers, one on each wing. Only one of the two should be considered since the method to compute wing mass based on one half and doubles it. This will be corrected later by introducing a symmetrical tag in the powertrain file which will allow the user to define components "pair" to avoid this kind of situation. The function will still output the whole list but it will also output said symmetrical list and the assemblers will do the sorting out (may need the total list for future use)