Symmetricality added
It is now possible to define a symmetrical component in the power train file, this way only one of the pair will be considered. The list of pairs is an output of the function that list the punctual loads and distributed loads and only the pairs of the specified type should appear.