Punctual masses assembler
Created an tested an assembler that creates a list of puncutal mases on the wing with the right format based on the data from the powertrain sizing. Same will be done for distributed masses but at a bigger scale.
Also a feature that check pairs when you declare them in the pt file to ensure that you pair it with something that actually exists