Wing sizing speed-up
Realized there was a way to speed up the wing sizing. Since only propulsor and propulsive loads are needed, we now produce a simplified pt file which contain only those and run the wing sizing with that. This file is then deleted at the end of the wing sizing so no overclogging