Initial guesses displacement
Moved the initial guesses to the top-most component so that we have access to all the input and thus make a right initial guess for the current. It seems to, for some reasons that I can only theorize, reduce the time it takes for computation. So good news all in all. Next step is to produce the initial guess for the current.