Initial guess econ
Added two function to get and set an initial guess for the thrust and airspeed which will be relevant for the energy consumption. Also realized something. This guess non linear function is ran before any module has time to run, meaning everything that is not a problem inputs should not be used !!! Or at least we should only use variable for which we will provide a good initial guess at problem level !!!