First attempt power at each node
Made a first attempt at setting the power at each node of the power train. Even though certain things might be done faster (think for loops in for loops) it works well and is fairly fast. This version simply works with setting the computation priority of components because this is how I originally imagine things would go. However, it might just as well work with power, which is what we will do in the next commit