Component syntax test
Created some tests to automatically verify that the component that are added in the list output what we expect of them. For instance, we ensure that all component sizing  output mass, drag, CGs slipstream outputs delta_Cl, Cd and Cm. For performance it is a bit trickier but we ensure that all TYPES of component output the right thing.

Also added this test in the workflow