Fuel tank performances
This component does not do much, its ultimate goal being to provide the fuel remaining in the tank at each step of the flight so that we can use it to compute the "varying" CG of the power train. The unconsumable part of the fuel will not appear in the fuel remaiing rather it will be included in the tank weight. This allows to have it inside the OWE and in the computation of the CG, in contrast with the previous solution.