Fuel tank slipstream
Does not do much expect output zero as expected. Next commit will be about plugging the tanks inside the pt builder like a generic component. Then we will add the assembler that properly compute the variation of CG and the "distributed" fuel loads (which will bring some slight change to the sizing component because we will need the y_ratios even for the pod because of how i was previously implemented)