Fuel System standalone
Implemented the component necessary for a standalone computation of the fuel system. For the performance side, only serve as a fuel distributor among tanks
For the sizing side, we will make the not so farfetched assumption the volume of connected tank is linked to the amount of fuel that flowed through the system during the whole mission.