Tests on how residuals work
The title says it all, I took a simplified problem and played a little with it to see how it worked. The result is that when computing resdiuals, OM first concatenante all output value into one bbig vector and takes the norm of that. Meaning that vector weigh way more than float output