Backpedal and toying around
Toyed around with the residuals on a classic case of retrofitting.  Realized that there are some residuals (which are otherwise irrelevant like the reluctance) which trumps all the others. Also realized that when choosing a unit that makes them smaller, it may take one or two more iteration to run. There may be a way to play with that !