Assembly with pt file
Made a special arrangement to the pt file so that you don't have to specify through an option when you directly connect a battery to a bus. For each time the battery is a target in a connection, we check if it is to a bus. If that's the case we enable the option on the battery performances component and reverse the OpenMDAO inputs/outputs