Mission test passing
The diret battery to bus connection has been tested in the mission unit tests, ensuring it work when sizing is coupled. There is one point of caution however, since the dc_current now become an output and the voltage out an input, we have to slightly alter the performances watcher. For now, only one component will be affected so I didn't bother thinking of a generic way to do it and we do the corretion directly in the function that creates the list of value to include in the performances watchers