Pipistrel validation
Created a new folder for the purpose of creating a ready-to-go test case. Here the Pipistrel SW128 was chosen. Data from the flight manual were used as well as data from the manufacturer's website. Pretty good fit all things considered. The goal of next commit will be to develop graph to illustrate that
Also small name correction